[Update] Test configuration now 'extends' base.
