+++ src/components/title/test/mount-spec.js
@@ -1,17 +1,23 @@
-/* eslint no-unused-expressions: 0 */
-import testHelper from 'helpers/test-helper';
-import App from 'components/app';
+import { expect } from 'helpers/test-helper';
 
-describe('Title', () => {
-  let component;
+import App, { styles } from 'components/app';
+import Title from 'components/title';
 
-  console.log(testHelper);
+describe('#Initial tests setup', function() {
+  it('-Render App component', function() {
+    const container = document.getElementById('app'),
+      _node = renderComponent(App, container, {}),
+      className = styles.app;
 
-  beforeEach(() => {
-    component = renderComponent(App);
+    expect(_node.className).to.equal(className);
   });
 
-  it('Renders', () => {
-    expect(component.find('.title')).to.exist;
+  it('-Render Title', function() {
+    const container = document.getElementById('app'),
+      _node = renderComponent(Title, container, {
+        title: 'App Title'
+      });
+
+    expect(_node.innerText).to.equal('App Title');
   });
 });
